Move win32 mman plug-in replacement code to separate files
